adsAsynPortDriver.cpp: Improve logging when adsReadStateLock fails
There was not really much information when adsReadStateLock() failed.
One case was that the crate was run with a test-license, which had expired.
Improve this a little bit by printing the error code.

Keep the loop to disconnect, reconnect and try again, to be able to recover
some day.
Make it less spammy by adding a sleep(10 seconds).